move CompanionDeviceManager back to SettingsFragment
